# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RentKarRideKar API Gateway",
    "description": "Complete interactive API documentation for RentKarRideKar services including Auth, Vehicle, KYC, Order, Earnings, Referral, Wallet, Disputes, Ratings, Communication, and Admin services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/vehicles/search": {
      "get": {
        "summary": "Search Available Vehicles",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vehicleType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seats",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "color",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ac",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
...
    "/vehicles/admin/all": {
      "get": {
        "summary": "List All Vehicles (Admin)",
        "parameters": [
          {
            "name": "ownerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles": {
      "post": {
        "summary": "Register New Vehicle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vehicleCategory": {
                    "type": "string"
                  },
                  "make": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  },
                  "variant": {
                    "type": "string"
                  },
                  "registrationNumber": {
                    "type": "string"
                  },
                  "seatingCapacity": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "hasAC": {
                    "type": "boolean"
                  },
                  "manufacturingYear": {
                    "type": "integer"
                  },
                  "serviceRadius": {
                    "type": "integer"
                  },
                  "homeLat": {
                    "type": "number"
                  },
                  "homeLng": {
                    "type": "number"
                  },
                  "homeAddress": {
                    "type": "string"
                  },
                  "vehiclePhotos": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created successfully"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/{id}": {
      "put": {
        "summary": "Update Vehicle Details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/{id}/block": {
      "post": {
        "summary": "Block Vehicle Custom Dates",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": {
                    "type": "string"
                  },
                  "endDate": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/blocks/{blockId}": {
      "delete": {
        "summary": "Unblock Vehicle Schedule",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/{id}/available": {
      "get": {
        "summary": "Check Vehicle Availability",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/{vehicle_id}/plate-type": {
      "patch": {
        "summary": "Update Vehicle Plate Type (NEW 1.1)",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plateType": {
                    "type": "string",
                    "enum": [
                      "WHITE",
                      "YELLOW"
                    ]
                  },
                  "commercialPermitNumber": {
                    "type": "string"
                  },
                  "permitType": {
                    "type": "string"
                  },
                  "permitExpiryDate": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "plateType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
    "/vehicles/{vehicle_id}/pricing": {
      "get": {
        "summary": "Get Vehicle Pricing Detail (NEW 1.3)",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      },
      "put": {
        "summary": "Set / Update Vehicle Pricing (NEW 1.4)",
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "perKmOutstation": {
                    "type": "number"
                  },
                  "perHourLocal": {
                    "type": "number"
                  },
                  "minimumBookingHours": {
                    "type": "integer"
                  },
                  "nightChargePercentage": {
                    "type": "number"
                  },
                  "eventPackage": {
                    "type": "object",
                    "properties": {
                      "halfDay": {
                        "type": "number"
                      },
                      "fullDay": {
                        "type": "number"
                      },
                      "weddingPackage": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Vehicle Service"
        ]
      }
    },
...
  }
}